+++ src/main/java/de/maxhenkel/easyvillagers/blocks/tileentity/VillagerTileentity.java
@@ -49,6 +49,8 @@ public class VillagerTileentity extends FakeWorldTileentity {
     public void setVillager(ItemStack villager) {
         this.villager = villager;
 
+        removeTradingPlayer();
+
         if (villager.isEmpty()) {
             villagerEntity = null;
         } else {
@@ -57,6 +59,12 @@ public class VillagerTileentity extends FakeWorldTileentity {
         }
         setChanged();
         sync();
+    }
+
+    public void removeTradingPlayer() {
+        if (villagerEntity != null) {
+            villagerEntity.setTradingPlayer(null);
+        }
     }
 
     protected void onAddVillager(EasyVillagerEntity villager) {
@@ -89,6 +97,12 @@ public class VillagerTileentity extends FakeWorldTileentity {
 
     public static boolean advanceAge(EasyVillagerEntity villagerEntity) {
         return advanceAge(villagerEntity, 1);
+    }
+
+    @Override
+    public void setRemoved() {
+        removeTradingPlayer();
+        super.setRemoved();
     }
 
     @Override